# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Dynamic User Segmentation Service.",
        "title": "Dynamic User Segmentation Service",
        "contact": {
            "name": "Keril",
            "email": "khe.14@yandex.ru"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/files/": {
            "get": {
                "description": "Get link to records",
                "tags": [
                    "record"
                ],
                "summary": "Get link to records",
                "responses": {
                    "301": {
                        "description": "Moved Permanently",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/records/": {
            "get": {
                "description": "Get link to records by year-month",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "record"
                ],
                "summary": "Get record by year-month",
                "parameters": [
                    {
                        "description": "date year-month",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.GetRecordsByYM.Request"
                        }
                    }
                ],
                "responses": {
                    "301": {
                        "description": "Moved Permanently"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/slugs/": {
            "post": {
                "description": "Delete slug",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "slugs"
                ],
                "summary": "Delete slug",
                "parameters": [
                    {
                        "description": "input title",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.DeleteSlug.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.DeleteSlug.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/slugsUsers/": {
            "get": {
                "description": "Get users slugs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Segments"
                ],
                "summary": "Get users slugs",
                "parameters": [
                    {
                        "description": "input id",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AddUserInSlugs.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GetUsersSlugs.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "Add users in slugs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Segments"
                ],
                "summary": "Add users in slugs",
                "parameters": [
                    {
                        "description": "input [title_add], [title_delete], id, ttl",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AddUserInSlugs.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.AddUserInSlugs.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/": {
            "post": {
                "description": "Create user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Create user",
                "parameters": [
                    {
                        "description": "input id",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateUser.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Users"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Slugs": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "entity.Users": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "handlers.AddUserInSlugs.Request": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title_add": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title_delete": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ttl": {
                    "type": "integer"
                }
            }
        },
        "handlers.AddUserInSlugs.Response": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "handlers.CreateSlug.Request": {
            "type": "object",
            "properties": {
                "part": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "handlers.CreateUser.Request": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "handlers.DeleteSlug.Request": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                }
            }
        },
        "handlers.DeleteSlug.Response": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "handlers.GetRecordsByYM.Request": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                }
            }
        },
        "handlers.GetUsersSlugs.Response": {
            "type": "object",
            "properties": {
                "slugs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    }
}